GPT-Oracle Lite v2.0.0/requirements.txt
discord.py==2.3.0
python-dotenv==1.0.0
openai==0.27.8
tiktoken==0.4.0
tokenizer==3.4.2
transformers==4.30.2
discord-py-slash-command==4.2.1
pyodbc==4.0.39
mysql-connector-python==8.0.33
discord-ext-bot==1.0.1
pymysql==1.0.3
asyncio==3.4.3
futures==3.4.0
